refactor tokenize script
