+++ plugins/dates/src/03-ranges/01-two-times.js
@@ -1,4 +1,4 @@
-const parseDate = require('../03-parse')
+const parseDate = require('../04-parse')
 const reverseMaybe = require('./_reverse')
 
 const moveToPM = function (obj) {
