+++ plugins/dates/src/03-ranges/02-date-range.js
@@ -1,4 +1,4 @@
-const parseDate = require('../03-parse')
+const parseDate = require('../04-parse')
 const reverseMaybe = require('./_reverse')
 
 module.exports = [
