+++ plugins/dates/src/03-ranges/03-one-date.js
@@ -1,6 +1,6 @@
-const parseDate = require('../03-parse')
+const parseDate = require('../04-parse')
 const reverseMaybe = require('./_reverse')
-const Unit = require('../03-parse/units/Unit')
+const Unit = require('../04-parse/units/Unit')
 
 const punt = function (unit, context) {
   unit = unit.applyShift(context.punt)
