+++ plugins/dates/src/04-parse/01-tokenize/07-weekday.js
