--- plugins/dates/tmp.html
@@ -1,102 +1,0 @@
-<html>
-
-<style>
-  body{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    align-items: center;
-    text-align:center;
-    flex-wrap: wrap;
-    align-self: stretch;
-  }
-  .container{
-    margin-top:200px;
-    color:#fbfbfb;
-    font-size:40px;
-    position:relative;
-  }
-  .punt{
-    position:relative;
-    padding:5px;
-    background-color:#2D85A8;
-  }
-  .relative{
-    position:relative;
-    padding:5px;
-    background-color:#735873;
-  }
-  .main{
-    position:relative;
-    padding:5px;
-    background-color:#F2C0BB;
-  }
-  .time{
-    position:relative;
-    padding:5px;
-    background-color:#978BA3;
-  }
-  .tz{
-    position:relative;
-    padding:5px;
-    background-color:#838B91;
-  }
-  .this{
-    position:relative;
-    padding:5px;
-    background-color:#AB5850;
-  }
-  .above{
-    position:absolute;
-    font-size:14px;
-    padding:3px;
-    top:-30px;
-    right:20px;
-    color:#fbfbfb;
-  }
-  .blue{
-    background-color:#2D85A8;
-  }
-  .purple{
-    background-color:#735873;
-  }
-  .pink{
-    background-color:#F2C0BB;
-  }
-  .fuscia{
-    background-color:#978BA3;
-  }
-  .grey{
-    background-color:#838B91;
-  }
-  .red{
-    background-color:#AB5850;
-  }
-</style>
-<body>
-
-<div class="container">
-
-<span class="punt">[two weeks and 4.5 days after]
-  <span class="above blue">punt</span>
-</span>
-<span class="relative">[the 2nd monday of]
-  <span class="above purple">ordinal</span>
-</span>
-<span class="this">[this]
-  <span class="above red">relative</span>
-</span>
-<span class="main">[feb]
-  <span class="above pink">root</span>
-</span>
-<span class="time">[at 2pm]
-  <span class="above fuscia">time</span>
-</span>
-<span class="tz">[kabul time]
-  <span class="above grey">timezone</span>
-</spa>
-
-</div>
-
-</body>
-</html>
